requirements.txt
Flask==3.0.3
flask-cors==5.0.0
gunicorn==22.0.0
google-cloud-storage==2.18.2
google-cloud-bigquery==3.25.0
google-auth==2.35.0
google-auth-oauthlib==1.2.1
python-dotenv==1.0.1
